+++ .dotnet/examples/Chat/Example07_StructuredOutputs.cs
@@ -42,7 +42,7 @@ public partial class ChatExamples
         };
 
         ChatCompletion chatCompletion = client.CompleteChat(
-            ["How can I solve 8x + 7 = -23?"],
+            [ new UserChatMessage("How can I solve 8x + 7 = -23?") ],
             options);
 
         using JsonDocument structuredJson = JsonDocument.Parse(chatCompletion.ToString());