+++ .dotnet/tests/Chat/OpenAIChatModelFactoryTests.cs
@@ -76,8 +76,8 @@ public partial class OpenAIChatModelFactoryTests
     public void ChatCompletionWithContentWorks()
     {
         IEnumerable<ChatMessageContentPart> content = [
-            ChatMessageContentPart.CreateTextMessageContentPart("first part"),
-            ChatMessageContentPart.CreateTextMessageContentPart("second part")
+            ChatMessageContentPart.CreateTextPart("first part"),
+            ChatMessageContentPart.CreateTextPart("second part")
         ];
         ChatCompletion chatCompletion = OpenAIChatModelFactory.ChatCompletion(content: content);
 
@@ -508,8 +508,8 @@ public partial class OpenAIChatModelFactoryTests
     public void StreamingChatCompletionUpdateWithContentUpdateWorks()
     {
         IEnumerable<ChatMessageContentPart> contentUpdate = [
-            ChatMessageContentPart.CreateTextMessageContentPart("first part"),
-            ChatMessageContentPart.CreateTextMessageContentPart("second part")
+            ChatMessageContentPart.CreateTextPart("first part"),
+            ChatMessageContentPart.CreateTextPart("second part")
         ];
         StreamingChatCompletionUpdate streamingChatCompletionUpdate = OpenAIChatModelFactory.StreamingChatCompletionUpdate(contentUpdate: contentUpdate);