+++ .dotnet/CHANGELOG.md
@@ -2,9 +2,21 @@
 
 ## 2.0.0-beta.12 (Unreleased)
 
+### Features Added
+
+### Breaking Changes
+
+- Renamed `ChatMessageContentPart`'s `CreateTextMessageContentPart` factory method to `CreateTextPart`.
+- Renamed `ChatMessageContentPart`'s `CreateImageMessageContentPart` factory method to `CreateImagePart`.
+- Renamed `ChatMessageContentPart`'s `CreateRefusalMessageContentPart` factory method to `CreateRefusalPart`.
+- Renamed `ImageChatMessageContentPartDetail` to `ChatImageDetailLevel`.
+- Removed `ChatMessageContentPart`'s `ToString` overload.
+
 ### Bugs Fixed
 
 - Addressed an issue that caused multi-page queries of fine-tuning jobs, checkpoints, and events to fail. (commit_hash)
+
+### Other Changes
 
 ## 2.0.0-beta.11 (2024-09-03)